refactor: GachaMachine3D에서 이징, 애니메이터, 가중치 계산 로직 hooks, utils 사용
