+++ src/utils/Easing.ts
@@ -1,0 +1,3 @@
+export const easeOutCubic = (t: number) => 1 - Math.pow(1 - t, 3);
+export const easeInOutCubic = (t: number) =>
+  t < 0.5 ? 4 * t * t * t : 1 - Math.pow(-2 * t + 2, 3) / 2;
